+++ android/app/src/main/kotlin/com/cnxin/lifelog_flutter_demo/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.cnxin.lifelog
+package com.cnxin.lifelog.v2
 
 import io.flutter.embedding.android.FlutterActivity